Additional refactoring and integration test
