+++ Cargo.toml
@@ -25,6 +25,7 @@ aws_lambda = { git = "https://github.com/srijs/rust-aws-lambda" }
 
 [dev-dependencies]
 rand = "0.5"
+approx = "0.2.0"
 
 [badges]
 travis-ci = { repository = "phillyfan1138/option_price_faas" }
