--- lambda/constants/keys.js
@@ -1,18 +1,0 @@
-const keys={
-    cgmy:0,
-    merton:1,
-    heston:2,
-    market:3,
-    putprice:0,
-    callprice:1,
-    putdelta:2,
-    calldelta:3,
-    putgamma:4,
-    callgamma:5,
-    puttheta:6,
-    calltheta:7,
-    density:8,
-    riskmetric:9
-}
-
-module.exports=keys
