+++ tests/parameter4.json
@@ -1,0 +1,4 @@
+{
+    "body":"{\"num_u\":8,\"rate\":0.1,\"maturity\":0.5,\"asset\":38,\"cf_parameters\":{\"sigma\":0.5,\"speed\":0.1,\"v0\":0.2,\"eta_v\":0.1,\"rho\":-0.5},\"strikes\":[100],\"quantile\":0.01}",
+    "path":"/v1/heston/riskmetric"
+}
